+++ source/user-base-auth/src/main/java/com/base/auth/controller/TaskQuestionController.java
@@ -13,11 +13,13 @@ import com.base.auth.form.taskQuestion.CreateTaskQuestionForm;
 import com.base.auth.form.taskQuestion.UpdateQuestionTaskForm;
 import com.base.auth.mapper.TaskQuestionMapper;
 import com.base.auth.model.Simulation;
+import com.base.auth.model.StudentTaskQuestionProgress;
 import com.base.auth.model.SubTask;
 import com.base.auth.model.Task;
 import com.base.auth.model.TaskQuestion;
 import com.base.auth.model.criteria.TaskQuestionCriteria;
 import com.base.auth.repository.SimulationRepository;
+import com.base.auth.repository.StudentTaskQuestionProgressRepository;
 import com.base.auth.repository.SubTaskRepository;
 import com.base.auth.repository.TaskQuestionRepository;
 import com.base.auth.repository.TaskRepository;
@@ -61,6 +63,9 @@ public class TaskQuestionController extends ABasicController{
   @Autowired
   SimulationRepository simulationRepository;
 
+  @Autowired
+  StudentTaskQuestionProgressRepository studentTaskQuestionProgressRepository;
+
   @PostMapping(value = "/create", produces = MediaType.APPLICATION_JSON_VALUE)
   @PreAuthorize("hasRole('TQ_C')")
   public ApiMessageDto<String> create(@Valid @RequestBody CreateTaskQuestionForm createTaskQuestionForm, BindingResult bindingResult){
@@ -68,17 +73,35 @@ public class TaskQuestionController extends ABasicController{
     if (!isEducator()){
       throw new BadRequestException("User is not an educator", ErrorCode.USER_ERROR_NOT_EDUCATOR);
     }
-    TaskQuestion taskQuestion = taskQuestionRepository.findByQuestion(createTaskQuestionForm.getQuestion()).orElse(null);
-    if (taskQuestion != null && Objects.equals(taskQuestion.getSubTask().getId(), createTaskQuestionForm.getSubTaskId())){
+    TaskQuestion taskQuestion = taskQuestionRepository.findByQuestionAndSubTaskId(createTaskQuestionForm.getQuestion(), createTaskQuestionForm.getSubTaskId()).orElse(null);
+    if (taskQuestion != null){
+      if (!validateQuestionType(taskQuestion.getQuestionType(), createTaskQuestionForm.getQuestionType())){
+        throw new BadRequestException("Question cannot be created due to a type conflict", ErrorCode.TASK_QUESTION_ERROR_NOT_CREATE);
+      }
       if (!Objects.equals(createTaskQuestionForm.getQuestionType(), UserBaseConstant.QUESTION_TYPE_MULTIPLE_CHOICE)){
         if (createTaskQuestionForm.getOptions() != null){
-          throw new BadRequestException("Options cannot be created", ErrorCode.TASK_QUESTION_ERROR_NOT_CREATE);
+          throw new BadRequestException("Options cannot be created", ErrorCode.TASK_QUESTION_ERROR_NOT_CREATE_OPTION);
         }
         throw new BadRequestException("Task question already exist", ErrorCode.TASK_QUESTION_ERROR_EXIST);
       } else {
         TaskQuestion existTaskQuestion = taskQuestionRepository.findByOptions(createTaskQuestionForm.getOptions()).orElse(null);
         if (existTaskQuestion != null){
           throw new BadRequestException("Task question already exist", ErrorCode.TASK_QUESTION_ERROR_EXIST);
+        }
+      }
+    } else{
+      TaskQuestion existTaskQuestionBySubTaskId = taskQuestionRepository.findFirstBySubTaskId(createTaskQuestionForm.getSubTaskId());
+      if (existTaskQuestionBySubTaskId != null && !validateQuestionType(existTaskQuestionBySubTaskId.getQuestionType(), createTaskQuestionForm.getQuestionType())){
+        throw new BadRequestException("Question cannot be created due to a type conflict", ErrorCode.TASK_QUESTION_ERROR_NOT_CREATE);
+      } else{
+        if (!Objects.equals(createTaskQuestionForm.getQuestionType(), UserBaseConstant.QUESTION_TYPE_MULTIPLE_CHOICE)){
+          if (createTaskQuestionForm.getOptions() != null){
+            throw new BadRequestException("Options cannot be created", ErrorCode.TASK_QUESTION_ERROR_NOT_CREATE_OPTION);
+          }
+        } else {
+          if (createTaskQuestionForm.getOptions() == null){
+            throw new BadRequestException("Options cannot be null", ErrorCode.TASK_QUESTION_ERROR_OPTION_NOT_NULL);
+          }
         }
       }
     }
@@ -92,6 +115,13 @@ public class TaskQuestionController extends ABasicController{
     taskQuestion = taskQuestionMapper.fromCreateTaskQuestionFormToEntity(createTaskQuestionForm);
     taskQuestion.setSubTask(subTask);
     taskQuestionRepository.save(taskQuestion);
+
+    if (Objects.equals(createTaskQuestionForm.getQuestionType(), UserBaseConstant.QUESTION_TYPE_MULTIPLE_CHOICE)){
+      int currentTotalQuestion = subTask.getTotalQuestion() + 1;
+      subTask.setTotalQuestion(currentTotalQuestion);
+      subTask.setMaxErrors((int) Math.ceil((double) currentTotalQuestion / 2));
+      subTaskRepository.save(subTask);
+    }
 
     if (Objects.equals(simulation.getStatus(), UserBaseConstant.STATUS_ACTIVE)){
       simulation.setStatus(UserBaseConstant.STATUS_WAITING_APPROVE);
@@ -159,23 +189,30 @@ public class TaskQuestionController extends ABasicController{
     TaskQuestion taskQuestion = taskQuestionRepository.findById(updateQuestionTaskForm.getId()).orElseThrow(()
     -> new NotFoundException("Task question not found", ErrorCode.TASK_QUESTION_ERROR_NOT_FOUND));
     if (!Objects.equals(taskQuestion.getQuestion(), updateQuestionTaskForm.getQuestion()) && Objects.equals(taskQuestion.getSubTask().getId(), updateQuestionTaskForm.getSubtaskId())){
-      if (!Objects.equals(updateQuestionTaskForm.getQuestionType(), UserBaseConstant.QUESTION_TYPE_MULTIPLE_CHOICE)){
-        TaskQuestion existTaskQuestion = taskQuestionRepository.findByQuestion(updateQuestionTaskForm.getQuestion()).orElse(null);
-        if (existTaskQuestion != null){
-          if (updateQuestionTaskForm.getOptions() != null){
-            throw new BadRequestException("Options cannot be update", ErrorCode.TASK_QUESTION_ERROR_NOT_UPDATE);
+      if (Objects.equals(taskQuestion.getQuestionType(), UserBaseConstant.QUESTION_TYPE_MULTIPLE_CHOICE)){
+        if (updateQuestionTaskForm.getOptions() == null){
+          throw new BadRequestException("Options cannot be null", ErrorCode.TASK_QUESTION_ERROR_OPTION_NOT_NULL);
+        } else {
+          TaskQuestion existTaskQuestion = taskQuestionRepository.findByQuestionAndOptions(updateQuestionTaskForm.getQuestion(), updateQuestionTaskForm.getOptions()).orElse(null);
+          if (existTaskQuestion != null){
+            throw new BadRequestException("Task question already exist", ErrorCode.TASK_QUESTION_ERROR_EXIST);
           }
-          throw new BadRequestException("Task question already exist", ErrorCode.TASK_QUESTION_ERROR_EXIST);
         }
       } else {
-        TaskQuestion existTaskQuestion = taskQuestionRepository.findByOptions(updateQuestionTaskForm.getOptions()).orElse(null);
-        if (existTaskQuestion != null){
-          throw new BadRequestException("Task question already exist", ErrorCode.TASK_QUESTION_ERROR_EXIST);
+        if (updateQuestionTaskForm.getOptions() != null){
+          throw new BadRequestException("Question cannot be updated", ErrorCode.TASK_QUESTION_ERROR_NOT_UPDATE);
+        } else {
+          TaskQuestion existTaskQuestion = taskQuestionRepository.findByQuestionAndSubTaskId(updateQuestionTaskForm.getQuestion(), updateQuestionTaskForm.getSubtaskId()).orElse(null);
+          if (existTaskQuestion != null){
+            throw new BadRequestException("Task question already exist", ErrorCode.TASK_QUESTION_ERROR_EXIST);
+          }
         }
       }
-    }
-    if (Objects.equals(taskQuestion.getQuestionType(), UserBaseConstant.QUESTION_TYPE_FILE) || Objects.equals(taskQuestion.getQuestionType(), UserBaseConstant.QUESTION_TYPE_TEXT)){
-      taskQuestion.setOptions(null);
+    } else if (!Objects.equals(taskQuestion.getSubTask().getId(), updateQuestionTaskForm.getSubtaskId())) {
+      TaskQuestion existTaskQuestionBySubTaskId = taskQuestionRepository.findFirstBySubTaskId(updateQuestionTaskForm.getSubtaskId());
+      if (existTaskQuestionBySubTaskId != null && !validateQuestionType(taskQuestion.getQuestionType(), existTaskQuestionBySubTaskId.getQuestionType())){
+        throw new BadRequestException("Question cannot be created due to a type conflict", ErrorCode.TASK_QUESTION_ERROR_NOT_CREATE);
+      }
     }
     SubTask subTask = subTaskRepository.findById(updateQuestionTaskForm.getSubtaskId()).orElseThrow(()
         -> new NotFoundException("Subtask not found", ErrorCode.SUBTASK_ERROR_NOT_FOUND));
@@ -216,12 +253,40 @@ public class TaskQuestionController extends ABasicController{
     if (!Objects.equals(simulation.getEducator().getId(), getCurrentUser())){
       throw new BadRequestException("Simulation cannot be deleted", ErrorCode.SIMULATION_ERROR_NOT_AUTHORIZED);
     }
+
+    if (Objects.equals(taskQuestion.getQuestionType(), UserBaseConstant.QUESTION_TYPE_MULTIPLE_CHOICE)){
+      int currentTotalQuestion = subTask.getTotalQuestion() - 1;
+      subTask.setTotalQuestion(currentTotalQuestion);
+      subTask.setMaxErrors((int) Math.ceil((double) currentTotalQuestion / 2));
+      subTaskRepository.save(subTask);
+    }
+
+    StudentTaskQuestionProgress studentTaskQuestionProgress = studentTaskQuestionProgressRepository.findFirstByTaskQuestionId(id).orElse(null);
+    if (studentTaskQuestionProgress != null){
+      throw new BadRequestException("Task question cannot be deleted", ErrorCode.TASK_ERROR_NOT_DELETE);
+    }
+
     taskQuestionRepository.delete(taskQuestion);
+
+
     if (Objects.equals(simulation.getStatus(), UserBaseConstant.STATUS_ACTIVE)){
       simulation.setStatus(UserBaseConstant.STATUS_WAITING_APPROVE);
       simulationRepository.save(simulation);
     }
     apiMessageDto.setMessage("Delete success");
     return apiMessageDto;
+  }
+
+  private Boolean validateQuestionType(Integer taskQuestionType, Integer createTaskQuestionType){
+    if ((Objects.equals(taskQuestionType, UserBaseConstant.QUESTION_TYPE_FILE)
+        || Objects.equals(taskQuestionType, UserBaseConstant.QUESTION_TYPE_TEXT))
+        && Objects.equals(createTaskQuestionType, UserBaseConstant.QUESTION_TYPE_MULTIPLE_CHOICE)){
+      return false;
+    } else if (Objects.equals(taskQuestionType, UserBaseConstant.QUESTION_TYPE_MULTIPLE_CHOICE)
+        && (Objects.equals(createTaskQuestionType, UserBaseConstant.QUESTION_TYPE_FILE)
+        || Objects.equals(createTaskQuestionType, UserBaseConstant.QUESTION_TYPE_TEXT))){
+      return false;
+    }
+    return true;
   }
 }
